Python Annotation & Async

+++ python_async_comprehension/0-async_generator.py
@@ -7,7 +7,7 @@ import random
 from typing import Iterable
 
 
-async def async_generator() -> Iterable[int]:
+async def async_generator() -> Iterable[float]:
     """Yield a random number between 0 and 10"""
     for _ in range(10):
         await asyncio.sleep(1)
